student/requirement.txt
Django==5.1.4
djangorestframework==3.15.2
python-decouple==3.8
sqlparse==0.5.3